requirements.txt
pyyaml
numpy
scipy
fastavro
pyquaternion
